Various minor updates
* Worker config is now a struct instead of string
* Moved ac_utils into store folder instead of scheduler
* Changed execution priority from i64 to i32
* Add PlatformPropertyValue::Unknown

+++ cas/scheduler/action_messages.rs
@@ -50,7 +50,7 @@ pub struct ActionInfo {
     /// The properties rules that must be applied when finding a worker that can run this action.
     pub platform_properties: PlatformProperties,
     /// The priority of the action. Higher value means it should execute faster.
-    pub priority: i64,
+    pub priority: i32,
     /// When this action was created.
     pub insert_timestamp: SystemTime,
